+++ kernels/rmsnorm.cu
@@ -1,26 +1,24 @@
 #include "kernels.h"
+#include <cuda_bf16.h>
 
-// TODO: Implement efficient RMSNorm kernel
-// Key optimization: use warp-level reduction for sum-of-squares
-// Formula: out = x * weight / sqrt(mean(x^2) + eps)
-__global__ void rmsnorm_kernel(half* out, const half* x, const half* weight,
-                                int dim, float eps) {
-    // Each block handles one row (one token position)
+// RMSNorm: out = x * weight / sqrt(mean(x^2) + eps)
+// Each block handles one row (one token position).
+// Shared memory warp reduction for sum-of-squares.
+__global__ void rmsnorm_kernel(__nv_bfloat16* out, const __nv_bfloat16* x,
+                                const __nv_bfloat16* weight, int dim, float eps) {
     int row = blockIdx.x;
-    const half* x_row = x + row * dim;
-    half* o_row = out + row * dim;
+    const __nv_bfloat16* x_row = x + row * dim;
+    __nv_bfloat16* o_row = out + row * dim;
 
-    // 1. Compute sum of squares (shared memory reduction)
     extern __shared__ float sdata[];
     float local_ss = 0.0f;
     for (int i = threadIdx.x; i < dim; i += blockDim.x) {
-        float val = __half2float(x_row[i]);
+        float val = __bfloat162float(x_row[i]);
         local_ss += val * val;
     }
     sdata[threadIdx.x] = local_ss;
     __syncthreads();
 
-    // Warp reduction
     for (int s = blockDim.x / 2; s > 0; s >>= 1) {
         if (threadIdx.x < s) sdata[threadIdx.x] += sdata[threadIdx.x + s];
         __syncthreads();
@@ -28,15 +26,14 @@ __global__ void rmsnorm_kernel(half* out, const half* x, const half* weight,
 
     float rms = rsqrtf(sdata[0] / dim + eps);
 
-    // 2. Normalize and scale
     for (int i = threadIdx.x; i < dim; i += blockDim.x) {
-        float val = __half2float(x_row[i]) * rms * __half2float(weight[i]);
-        o_row[i] = __float2half(val);
+        float val = __bfloat162float(x_row[i]) * rms * __bfloat162float(weight[i]);
+        o_row[i] = __float2bfloat16(val);
     }
 }
 
 namespace kernels {
-void rmsnorm(half* out, const half* x, const half* weight,
+void rmsnorm(__nv_bfloat16* out, const __nv_bfloat16* x, const __nv_bfloat16* weight,
              int seq_len, int dim, float eps) {
     int threads = (dim < 1024) ? dim : 1024;
     rmsnorm_kernel<<<seq_len, threads, threads * sizeof(float)>>>(
