+++ kernels/rope.cu
@@ -4,7 +4,7 @@
 // TODO: Implement RoPE kernel
 // Apply rotary positional embedding to Q and K tensors
 // For each pair (x[2i], x[2i+1]): rotate by angle = pos * theta^(-2i/head_dim)
-__global__ void rope_kernel(half* q, half* k,
+__global__ void rope_kernel(__nv_bfloat16* q, __nv_bfloat16* k,
                              int seq_len, int num_heads, int num_kv_heads,
                              int head_dim, int start_pos, float theta) {
     // TODO: implement
@@ -12,7 +12,7 @@ __global__ void rope_kernel(half* q, half* k,
 }
 
 namespace kernels {
-void rope(half* q, half* k, int seq_len, int num_heads, int num_kv_heads,
+void rope(__nv_bfloat16* q, __nv_bfloat16* k, int seq_len, int num_heads, int num_kv_heads,
           int head_dim, int start_pos, float theta) {
     // TODO: Launch rope_kernel
 }
